+++ modulos/dipucrrules/src/main/java/es/dipucr/sigem/api/rule/procedures/planesProvinciales/DipucrGeneraAnunciosPlanes2.java
@@ -198,13 +198,22 @@ public class DipucrGeneraAnunciosPlanes2 implements IRule {
 		    		gendocAPI.setDocument(connectorSession, documentId, docref, in, (int)(fileOut.length()), mime);
 		    		
 		    		//Borra archivos temporales
-		    		file.delete();
-		    		fileOut.delete();		    		
+		    		if(null != file) {
+		    			file.delete();
+		    		}
+		    		if(null != fileOut) {
+		    			fileOut.delete();		    		
+		    		}
 
-		    		entityTemplateT.delete(cct);
-					entityDocumentT.delete(cct);
-					ooHelper.dispose();
-		    		
+		    		if(null != entityTemplateT) {
+		    			entityTemplateT.delete(cct);
+		    		}
+					if(null != entityDocumentT) {
+						entityDocumentT.delete(cct);
+					}
+					if(null != ooHelper){
+			        	ooHelper.dispose();
+					}		    		
 	        	}
     		}	
 			cct.endTX(true);
@@ -320,13 +329,24 @@ public class DipucrGeneraAnunciosPlanes2 implements IRule {
 		    		gendocAPI.setDocument(connectorSession, documentId, docref, in, (int)(fileOut.length()), mime);
 		    		
 		    		//Borra archivos temporales
-		    		file.delete();
-		    		fileOut.delete();
+		    		if( null != file) { 
+		    			file.delete();
+		    		}
+		    		if (null != fileOut) {
+		    			fileOut.delete();
+		    		}
 
-		    		entityTemplateT.delete(cct);
-					entityDocumentT.delete(cct);
-					ooHelper.dispose();
-	        	}
+		    		if( null != entityTemplateT) {
+		    			entityTemplateT.delete(cct);
+		    		}
+					if(null != entityDocumentT) {
+						entityDocumentT.delete(cct);
+					}
+
+					if(null != ooHelper){
+			        	ooHelper.dispose();
+					}
+				}
     		}	
 			cct.endTX(true);
 		} catch(Exception e) {
